@@ -1,11 +1,20 @@
 module oauth2-proxy-nexus3
 
-go 1.14
+go 1.19
 
 require (
-	github.com/caarlos0/env/v6 v6.2.1
-	github.com/google/uuid v1.1.1
-	github.com/gorilla/mux v1.7.4
-	github.com/stretchr/testify v1.5.1
-	github.com/thoas/go-funk v0.6.0
+	github.com/caarlos0/env/v8 v8.0.0
+	github.com/google/uuid v1.3.0
+	github.com/gorilla/mux v1.8.0
+	github.com/stretchr/testify v1.8.0
+	github.com/thoas/go-funk v0.9.3
+	go.uber.org/zap v1.24.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	go.uber.org/atomic v1.11.0 // indirect
+	go.uber.org/multierr v1.11.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
