@@ -1,4 +1,4 @@
-package gitlab
+package oicd_generic
 
 import (
 	"encoding/json"
@@ -9,8 +9,6 @@ import (
 	"oauth2-proxy-nexus3/authprovider"
 )
 
-const userInfoEndpointPath = "/oauth/userinfo"
-
 // Client implements `authprovider.Client`.
 type Client struct {
 	URL *url.URL
@@ -18,35 +16,35 @@ type Client struct {
 
 // GetUserInfo implements `authprovider.Client`.
 func (s *Client) GetUserInfo(accessToken string) (authprovider.UserInfo, error) {
-	endpoint, err := url.Parse(fmt.Sprintf(s.URL.String() + userInfoEndpointPath))
+	endpoint, err := url.Parse(s.URL.String())
 	if err != nil {
-		return nil, fmt.Errorf("Failed to parse the GitLab URL: %s", err)
+		return nil, fmt.Errorf("Failed to parse the OpenID Connect URL: %s", err)
 	}
 
 	req, err := http.NewRequest("GET", endpoint.String(), nil)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create the GitLab GET userinfo request: %s", err)
+		return nil, fmt.Errorf("failed to create the OpenID Connect GET userinfo request: %s", err)
 	}
 
 	req.Header.Set("Authorization", "Bearer "+accessToken)
 
 	res, err := http.DefaultClient.Do(req)
 	if err != nil {
-		return nil, fmt.Errorf("failed to request the GitLab GET userinfo endpoint on %s: %s", s.URL, err)
+		return nil, fmt.Errorf("failed to request the OpenID Connect GET userinfo endpoint on %s: %s", s.URL, err)
 	}
 	defer res.Body.Close()
 
 	if res.StatusCode != http.StatusOK {
 		if resBody, err := io.ReadAll(res.Body); err == nil {
-			return nil, fmt.Errorf("failed to request the GitLab GET userinfo endpoint on %s: %s", s.URL, resBody)
+			return nil, fmt.Errorf("failed to request the OpenID Connect GET userinfo endpoint on %s: %s", s.URL, resBody)
 		}
 
-		return nil, fmt.Errorf("failed to read the GitLab GET userinfo error response: %s", err)
+		return nil, fmt.Errorf("failed to read the OpenID Connect GET userinfo error response: %s", err)
 	}
 
 	var userInfo UserInfo
 	if err := json.NewDecoder(res.Body).Decode(&userInfo); err != nil {
-		return nil, fmt.Errorf("failed to decode the GitLab GET userinfo responses: %s", err)
+		return nil, fmt.Errorf("failed to decode the OpenID Connect GET userinfo responses: %s", err)
 	}
 
 	return &userInfo, nil
